third_party/blink/web_tests/svg/animations/animate-marker-orient-from-auto-to-angle-expected.txt

Animate SVGMarkerElement orientAttr from auto to an angle

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS marker.orientAngle.animVal.value is 0
PASS marker.orientAngle.baseVal.value is 0
PASS marker.orientType.animVal is SVGMarkerElement.SVG_MARKER_ORIENT_ANGLE
PASS marker.orientType.baseVal is SVGMarkerElement.SVG_MARKER_ORIENT_ANGLE
PASS marker.orientAngle.animVal.value is 0
PASS marker.orientAngle.baseVal.value is 0
PASS marker.orientType.animVal is SVGMarkerElement.SVG_MARKER_ORIENT_AUTO
PASS marker.orientType.baseVal is SVGMarkerElement.SVG_MARKER_ORIENT_ANGLE
PASS marker.orientAngle.animVal.value is 0
PASS marker.orientAngle.baseVal.value is 0
PASS marker.orientType.animVal is SVGMarkerElement.SVG_MARKER_ORIENT_AUTO
PASS marker.orientType.baseVal is SVGMarkerElement.SVG_MARKER_ORIENT_ANGLE
PASS marker.orientAngle.animVal.value is 90
PASS marker.orientAngle.baseVal.value is 0
PASS marker.orientType.animVal is SVGMarkerElement.SVG_MARKER_ORIENT_ANGLE
PASS marker.orientType.baseVal is SVGMarkerElement.SVG_MARKER_ORIENT_ANGLE
PASS marker.orientAngle.animVal.value is 90
PASS marker.orientAngle.baseVal.value is 0
PASS marker.orientType.animVal is SVGMarkerElement.SVG_MARKER_ORIENT_ANGLE
PASS marker.orientType.baseVal is SVGMarkerElement.SVG_MARKER_ORIENT_ANGLE
PASS marker.orientAngle.animVal.value is 90
PASS marker.orientAngle.baseVal.value is 0
PASS marker.orientType.animVal is SVGMarkerElement.SVG_MARKER_ORIENT_ANGLE
PASS marker.orientType.baseVal is SVGMarkerElement.SVG_MARKER_ORIENT_ANGLE
PASS successfullyParsed is true

TEST COMPLETE

SVG 1.1 dynamic animation tests
